requirements.txt
streamlit
torch
torchvision
numpy
matplotlib
scikit-learn
seaborn
Pillow